+++ react/components/pages/Details/Payment/PaymentCard.tsx
@@ -1,15 +1,23 @@
 import React, { FunctionComponent } from 'react'
-import { FormattedMessage } from 'react-intl'
+import { FormattedMessage, defineMessages } from 'react-intl'
 import { Button } from 'vtex.styleguide'
 
-import {
-  SubscriptionStatusEnum,
-  CSS,
-  PAYMENT_DIV_ID,
-  TagTypeEnum,
-} from '../../../../constants'
+import { CSS, PAYMENT_DIV_ID, TagTypeEnum } from '../../../../constants'
 import Alert from '../../../commons/CustomAlert'
+import EditAlert from '../../../commons/EditAlert'
+import EditButton from '../../../commons/EditButton'
 import PaymentDisplay from '../PaymentDisplay'
+
+const messages = defineMessages({
+  label: {
+    id: 'subscription.purchase-settings.error.action',
+    defaultMessage: '',
+  },
+  noAction: {
+    id: 'subscription.purchase-settings.error.no-action',
+    defaultMessage: '',
+  },
+})
 
 const SubscriptionsGroupPaymentCard: FunctionComponent<Props> = ({
   subscriptionsGroup,
@@ -17,76 +25,58 @@ const SubscriptionsGroupPaymentCard: FunctionComponent<Props> = ({
   onMakeRetry,
   displayRetry,
   isRetryButtonEnabled,
-}) => {
-  // TODO: Check if Boleto is working
-  // const lastGeneratedOrder = subscriptionsGroup.lastInstance
-  // const bankInvoiceUrl =
-  //   lastGeneratedOrder &&
-  //   lastGeneratedOrder.orderInfo &&
-  //   lastGeneratedOrder.orderInfo.paymentUrl
-
-  const displayEdit =
-    subscriptionsGroup.status === SubscriptionStatusEnum.Active
-
-  return (
-    <div className={CSS.cardWrapper} id={PAYMENT_DIV_ID}>
-      <Alert
-        visible={displayRetry}
-        type={TagTypeEnum.Warning}
-        contentId="subscription.payment.alert.info.message"
-        onClose={() => null}
-      />
-      <div className="flex flex-row">
-        <div className="db-s di-ns b f4 tl c-on-base">
-          <FormattedMessage id="subscription.payment" />
-        </div>
-        <div className="ml-auto flex flex-row">
-          {displayRetry && (
-            <Button
-              size="small"
-              variation="secondary"
-              onClick={onMakeRetry}
-              disabled={!isRetryButtonEnabled}
-            >
-              <FormattedMessage id="subscription.retry.button.message" />
-            </Button>
-          )}
-          {displayEdit && (
-            <div className="ml3">
-              <Button size="small" variation="tertiary" onClick={onEdit}>
-                <FormattedMessage id="subscription.actions.edit" />
-              </Button>
-            </div>
-          )}
-        </div>
+}) => (
+  <div className={CSS.cardWrapper} id={PAYMENT_DIV_ID}>
+    <Alert
+      visible={displayRetry}
+      type={TagTypeEnum.Warning}
+      contentId="subscription.payment.alert.info.message"
+      onClose={() => null}
+    />
+    <div className="flex flex-row">
+      <div className="db-s di-ns b f4 tl c-on-base">
+        <FormattedMessage id="subscription.payment" />
       </div>
-      <div className="flex pt3-s pt0-ns w-100 mr-auto flex-row-ns flex-column-s">
-        <div className="f5-ns f6-s pt5 lh-solid dib-ns c-on-base">
-          <PaymentDisplay
-            purchaseSettings={subscriptionsGroup.purchaseSettings}
+      <div className="ml-auto flex flex-row">
+        {displayRetry && (
+          <Button
+            size="small"
+            variation="secondary"
+            onClick={onMakeRetry}
+            disabled={!isRetryButtonEnabled}
+          >
+            <FormattedMessage id="subscription.retry.button.message" />
+          </Button>
+        )}
+        <div className="ml3">
+          <EditButton
+            onEdit={onEdit}
+            subscriptionStatus={subscriptionsGroup.status}
+            testId="edit-payment-button"
           />
         </div>
-        {/* {subscriptionsGroup.purchaseSettings.paymentMethod
-          .paymentSystemGroup === 'bankInvoice' &&
-          bankInvoiceUrl && (
-            <div className="pl9-ns pt2-ns pt6-s">
-              <Button
-                block
-                size="small"
-                onClick={() => handleInvoiceButtonClick(bankInvoiceUrl)}
-                variation="secondary">
-                <FormattedMessage id="subscription.payment.invoice" />
-              </Button>
-            </div>
-          )} */}
       </div>
     </div>
-  )
-}
-
-// function handleInvoiceButtonClick(bankInvoiceUrl: string) {
-//   window.open(bankInvoiceUrl, '_blank')
-// }
+    <div className="flex pt3-s pt0-ns w-100 mr-auto flex-row-ns flex-column-s">
+      <div className="f5-ns f6-s pt5 lh-solid dib-ns c-on-base">
+        {subscriptionsGroup.purchaseSettings.paymentMethod ? (
+          <PaymentDisplay
+            paymentMethod={subscriptionsGroup.purchaseSettings.paymentMethod}
+          />
+        ) : (
+          <EditAlert
+            subscriptionStatus={subscriptionsGroup.status}
+            onAction={onEdit}
+            actionLabelMessage={messages.label}
+            noActionMessage={messages.noAction}
+          >
+            <FormattedMessage id="subscription.purchase-settings.error.message" />
+          </EditAlert>
+        )}
+      </div>
+    </div>
+  </div>
+)
 
 interface Props {
   subscriptionsGroup: SubscriptionsGroupItemType