+++ react/components/pages/Details/PaymentDisplay.tsx
@@ -3,38 +3,31 @@ import { InjectedIntlProps, injectIntl } from 'react-intl'
 
 import PaymentFlagIcon from '../../commons/PaymentFlagIcon'
 
-const PaymentDisplay: FunctionComponent<Props> = ({
-  purchaseSettings,
-  intl,
-}) => {
-  return purchaseSettings.paymentMethod.paymentAccount ? (
+const PaymentDisplay: FunctionComponent<Props> = ({ paymentMethod, intl }) => {
+  return paymentMethod.paymentAccount ? (
     <Fragment>
       <PaymentFlagIcon
-        group={purchaseSettings.paymentMethod.paymentSystemGroup}
-        type={purchaseSettings.paymentMethod.paymentSystemName}
+        group={paymentMethod.paymentSystemGroup}
+        type={paymentMethod.paymentSystemName}
         size={400}
       />
       <span className="fw3 f5-ns f6-s">
         {`${intl.formatMessage({
           id: 'subscription.payment.final',
-        })} ${purchaseSettings.paymentMethod.paymentAccount.cardNumber.slice(
-          -4
-        )}`}
+        })} ${paymentMethod.paymentAccount.cardNumber.slice(-4)}`}
       </span>
     </Fragment>
   ) : (
     <span className="fw3 f5-ns f6-s">
       {intl.formatMessage({
-        id: `paymentData.paymentGroup.${
-          purchaseSettings.paymentMethod.paymentSystemGroup
-        }.name`,
+        id: `paymentData.paymentGroup.${paymentMethod.paymentSystemGroup}.name`,
       })}
     </span>
   )
 }
 
 interface Props extends InjectedIntlProps {
-  purchaseSettings: PurchaseSettings
+  paymentMethod: PaymentMethod
 }
 
 export default injectIntl(PaymentDisplay)